+++ Packages/ai.fxn.fxn3d/Unity/FunctionUnity.cs
@@ -20,10 +20,7 @@ namespace Function {
 
     using System;
     using System.Collections.Generic;
-    using System.IO;
-    using System.Threading.Tasks;
     using UnityEngine;
-    using UnityEngine.Networking;
     using Unity.Collections.LowLevel.Unsafe;
     using API;
     using Internal;