+++ Changelog.md
@@ -1,4 +1,5 @@
 ## 0.0.34
++ Added `fxn.Beta.Predictions.Remote.Create` method for creating predictions that run on remote servers.
 + Removed `FunctionUnity.StreamingAssetsToAbsolutePath` method.
 
 ## 0.0.33